Moved post-update hook to the end of the update hook.
Removed the post-update hook from this branch. It can still be found
in a new branch created called 'using_post_update'. This change was
implemented because I don't want anything to be in the server's publish
branch that isn't actually posted to the database; post-update cannot
stop the saving to the repo, update can. There was a race condidition
between repo commit and database upload.